services/order-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
kafka-python==2.0.2
pydantic==2.5.3
pydantic-settings==2.1.0
prometheus-client==0.19.0
python-dotenv==1.0.0
